A fix for "per-agent verifier-issued epoch timestamp"
In some database backends, such as MySQL, the use of a float as column
lead to representation and rounding errors. In order to preserve the
information in the most backend-agnostic way possible, we will simply
use integer (nobody really cares about fractions of seconds for the
purpose of fault detection/"lagging behind").

+++ keylime/db/verifier_db.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Float, ForeignKey, Integer, LargeBinary, PickleType, String, Text, schema
+from sqlalchemy import Column, ForeignKey, Integer, LargeBinary, PickleType, String, Text, schema
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 
@@ -47,7 +47,7 @@ class VerfierMain(Base):
     ak_tpm = Column(String(500))
     mtls_cert = Column(String(2048), nullable=True)
     attestation_count = Column(Integer)
-    last_received_quote = Column(Float)
+    last_received_quote = Column(Integer)
     tpm_clockinfo = Column(JSONPickleType(pickler=JSONPickler))
