+++ keylime/migrations/versions/a09cc94177f0_add_last_received_quote.py
@@ -32,7 +32,7 @@ def downgrade_registrar():
 
 
 def upgrade_cloud_verifier():
-    op.add_column("verifiermain", sa.Column("last_received_quote", sa.Float, nullable=True))
+    op.add_column("verifiermain", sa.Column("last_received_quote", sa.Integer(), nullable=True))
 
 
 def downgrade_cloud_verifier():
